requirements.txt
Flask
pandas
matplotlib
fpdf
google-api-python-client
google-auth
gunicorn
